+++ gossip/hooks.py
@@ -1,8 +1,10 @@
 import functools
 import logbook
 import sys
+from collections import OrderedDict
 from contextlib import contextmanager
 from types import GeneratorType
+from sentinels import Sentinel
 
 from . import groups, registry
 from ._compat import itervalues, string_types
@@ -12,14 +14,16 @@ from .exceptions import (CannotResolveDependencies, HookNotFound,
 from .registration import Registration
 from .utils import topological_sort_registrations
 
-from logbook.utils import log_deprecation_message
+from vintage import warn_deprecation
 
 _logger = logbook.Logger(__name__)
+
+_REGISTER_NO_OP = Sentinel('REGISTER_NO_OP')
 
 
 class Hook(object):
 
-    def __init__(self, group, name, arg_names=(), doc=None, deprecated=False):
+    def __init__(self, group, name, arg_names=None, doc=None, deprecated=False):
         super(Hook, self).__init__()
         self.group = group
         self.name = name
@@ -30,13 +34,28 @@ class Hook(object):
             self.full_name = "{0}.{1}".format(self.group.full_name, self.name)
         registry.hooks[self.full_name] = self
         self._registrations = []
-        self._arg_names = arg_names
+        self._empty_regisrations = []
+        self._arguments = self._normalize_arguments(arg_names)
         self._trigger_internal_hooks = self.full_name != "gossip.on_handler_exception"
         self._defined = False
         self._pre_trigger_callbacks = []
         self._unmet_deps = frozenset()
         self.doc = doc
         self.deprecated = deprecated
+
+    def _get_registration_list_from_func(self, func):
+        if func is _REGISTER_NO_OP:
+            return self._empty_regisrations
+        return self._registrations
+
+    def _normalize_arguments(self, arg_names):
+        if arg_names is None:
+            return None
+
+        if isinstance(arg_names, dict):
+            return OrderedDict(arg_names)
+
+        return OrderedDict((arg_name, None) for arg_name in arg_names)
 
     def add_pre_trigger_callback(self, callback):
         self._pre_trigger_callbacks.append(callback)
@@ -64,7 +83,7 @@ class Hook(object):
         return self._defined
 
     def get_argument_names(self):
-        return self._arg_names
+        return tuple(self._arguments)
 
     def __call__(self, **kwargs):
         return self.trigger(kwargs)
@@ -80,12 +99,28 @@ class Hook(object):
             raise UnsupportedHookTags("hook {0} support {1} tags, not: {2}".format(
                 self.full_name, self.tags, extra_tags))
 
+    def validate_kwargs(self, kwargs):
+        if self._arguments is None or not self.group.is_strict():
+            return
+
+        unknown = set(kwargs) - set(self._arguments)
+        if unknown:
+            raise TypeError('Unknown arguments specified: {}'.format(', '.join(unknown)))
+
+        for arg_name, arg_types in self._arguments.items():
+            if arg_name not in kwargs:
+                raise TypeError('Missing argument {!r}'.format(arg_name))
+            if arg_types is not None and not isinstance(kwargs[arg_name], arg_types):
+                raise TypeError('Incorrect type for argument {}. Expected {!r}, got {!r}'.format(
+                    arg_name, arg_types, type(kwargs[arg_name])))
+
+
     def validate_strict(self, registrations_to_validate=None):
         if not self._defined:
             raise UndefinedHook(
                 "hook {0} wasn't defined yet".format(self.full_name))
         if registrations_to_validate is None:
-            registrations_to_validate = self._registrations
+            registrations_to_validate = self._registrations + self._empty_regisrations
         for registration in registrations_to_validate:
             self.validate_tags(registration.tags, is_strict=True)
 
@@ -93,40 +128,54 @@ class Hook(object):
         """Registers a new handler to this hook
         """
         if self.deprecated:
-            log_deprecation_message('Hook {0} is deprecated!'.format(self.full_name), frame_correction=+1)
-        returned = Registration(func, self, token=token, tags=tags, needs=needs, provides=provides, **kwargs)
+            warn_deprecation('Hook {0} is deprecated!'.format(self.full_name), frame_correction=+1)
+        new_registration = Registration(func, self, token=token, tags=tags, needs=needs, provides=provides, **kwargs)
         if self.group.is_strict():
-            self.validate_strict([returned])
-        self._registrations.append(returned)
-        if returned.needs or returned.provides:
+            self.validate_strict([new_registration])
+        need_sorting = self._registrations and new_registration.priority > self._registrations[-1].priority
+        self._get_registration_list_from_func(func).append(new_registration)
+        if need_sorting:
+            self._registrations.sort(key=Registration.get_priority, reverse=True) # sort is stable, so order among registration isn't disturbed
+        if new_registration.needs or new_registration.provides:
             try:
                 self.recompute_call_order()
             except:
                 self._registrations.pop()
                 raise
-        return returned
+        return new_registration
+
+    def register_no_op(self, **kwargs):
+        if kwargs.get('needs'):
+            raise NotImplementedError("Cannot define 'needs' for register_no_op")
+        return self.register(func=_REGISTER_NO_OP, **kwargs)
 
     def recompute_call_order(self):
         self._registrations = topological_sort_registrations(self._registrations, unconstrained_priority=self.group.get_unconstrained_handler_priority())
-        self._unmet_deps = frozenset(n for r in self._registrations for n in r.needs) - frozenset(p for r in self._registrations for p in r.provides)
+        self._unmet_deps = frozenset(n for r in self._registrations for n in r.needs) - \
+                           frozenset(p for r in (self._registrations + self._empty_regisrations) for p in r.provides)
 
 
     def unregister(self, registration):
         assert registration.hook is self
-        self._registrations.remove(registration)
+        self._get_registration_list_from_func(registration.func).remove(registration)
         registration.hook = None
+        self.recompute_call_order()
 
     def unregister_all(self):
         del self._registrations[:]
+        del self._empty_regisrations[:]
+        self.recompute_call_order()
 
     def trigger(self, kwargs, tags=None):
         if self._unmet_deps:
-            raise CannotResolveDependencies('Hook {0!r} has unmet dependencies: {1}'.format(self, ', '.join(map(str, self._unmet_deps))))
+            deps_str = ', '.join([str(dep) for dep in self._unmet_deps])
+            raise CannotResolveDependencies('Hook {0!r} has unmet dependencies: {1}'.format(self, deps_str))
         if self.full_name in _muted_stack[-1]:
             _logger.debug("Hook {0!r} muted, skipping trigger", self)
             return
 
         self.validate_tags(tags)
+        self.validate_kwargs(kwargs)
         exception_policy = self.group.get_exception_policy()
 
         registrations = self._registrations
@@ -166,21 +215,19 @@ class Hook(object):
         for callback in self._pre_trigger_callbacks:
             callback(registration, kwargs)
         try:
-            registration(**kwargs)  # pylint: disable=star-args
+            registration(**kwargs)
         except NotNowException:
             raise
-        except:
+        except Exception:  # pylint: disable=broad-except
             exc_info = sys.exc_info()
             if self._trigger_internal_hooks:
                 trigger("gossip.on_handler_exception",
                         handler=registration.func, exception=exc_info, hook=self)
-            _logger.debug("Exception occurred while calling {0}",
-                         registration, exc_info=exc_info)
-            # TODO: debug here
+            _logger.debug("Exception occurred while calling {0}", registration, exc_info=exc_info)
         return exc_info
 
     def __repr__(self):
-        return "<Hook {0}({1})>".format(self.name, ", ".join(self._arg_names))
+        return "<Hook {0}({1})>".format(self.name, ", ".join(self._arguments or ()))
 
 
 def trigger(hook_name, **kwargs):
@@ -255,7 +302,7 @@ def define(hook_name, **kwargs):
     return returned
 
 
-def register(func=None, hook_name=None, token=None, tags=None, needs=None, provides=None, reentrant=True, toggles_on=None, toggles_off=None):
+def register(func=None, hook_name=None, token=None, tags=None, needs=None, provides=None, reentrant=True, toggles_on=None, toggles_off=None, priority=0):
     """Registers a new function to a hook
 
     :param hook_name: full name of hook to register to
@@ -284,12 +331,16 @@ def register(func=None, hook_name=None, token=None, tags=None, needs=None, provi
             provides=provides,
             reentrant=reentrant,
             toggles_on=toggles_on,
-            toggles_off=toggles_off)
+            toggles_off=toggles_off,
+            priority=priority,
+        )
     assert hook_name is not None
     registration = get_or_create_hook(
         hook_name).register(
             func, token=token, tags=tags, needs=needs, provides=provides, reentrant=reentrant,
-            toggles_on=toggles_on, toggles_off=toggles_off)
+            toggles_on=toggles_on, toggles_off=toggles_off,
+            priority=priority,
+        )
     assert registration
     return func