+++ tests/test_needs_provides.py
@@ -3,11 +3,39 @@ import pytest
 import gossip
 from gossip.exceptions import CannotResolveDependencies
 
+
 def test_needs_provides_simple(timeline):
     evt1 = timeline.register(needs=['something'])
     evt2 = timeline.register(provides=['something'])
     timeline.trigger()
     assert evt1.timestamp > evt2.timestamp
+
+
+def test_unmet_deps_after_unregister(hook, counter):
+    reg = hook.register(counter.plus_one, provides=['something'])
+    hook.register(counter.plus_one, needs=['something'])
+
+    reg.unregister()
+    with pytest.raises(CannotResolveDependencies):
+        hook.trigger({})
+    assert counter.get() == 0
+
+
+def test_unmet_deps_resolved_after_unregister_all(hook, counter):
+    hook.register(counter.plus_one, needs=['1'])
+    hook.register(counter.plus_one, needs=['2'])
+    hook.register_no_op(provides=['3'])
+    assert len(hook._registrations) == 2  # pylint: disable=protected-access
+    assert len(hook._empty_regisrations) == 1  # pylint: disable=protected-access
+    assert len(hook._unmet_deps) == 2  # pylint: disable=protected-access
+
+    hook.unregister_all()
+    assert len(hook._registrations) == 0  # pylint: disable=protected-access
+    assert len(hook._empty_regisrations) == 0  # pylint: disable=protected-access
+    assert len(hook._unmet_deps) == 0  # pylint: disable=protected-access
+
+    hook.trigger({})
+    assert counter.get() == 0
 
 
 def test_needs_provides_complex(timeline):
@@ -31,7 +59,7 @@ def test_needs_provides_cyclic_simple(timeline):
     with pytest.raises(CannotResolveDependencies):
         timeline.register(needs=['0'], provides=['1'])
     # make sure the cyclic registration was not added
-    assert len(gossip.get_hook(timeline.hook_name)._registrations) == 1
+    assert len(gossip.get_hook(timeline.hook_name)._registrations) == 1  # pylint: disable=protected-access
 
 
 def test_needs_provides_cyclic_complex(timeline):
@@ -42,7 +70,42 @@ def test_needs_provides_cyclic_complex(timeline):
     with pytest.raises(CannotResolveDependencies):
         timeline.register(needs=['0'], provides=['4'])
     # make sure the cyclic registration was not added
-    assert len(gossip.get_hook(timeline.hook_name)._registrations) == 4
+    assert len(gossip.get_hook(timeline.hook_name)._registrations) == 4  # pylint: disable=protected-access
+
+
+def test_register_no_op_does_not_allow_needs_parameter(hook):
+    with pytest.raises(NotImplementedError):
+        hook.register_no_op(needs=['something'])
+    assert len(hook._unmet_deps) == 0  # pylint: disable=protected-access
+
+
+def test_register_no_op_solves_dependencies(timeline):
+    timeline.register(needs=['1'], provides=['0'])
+    timeline.register(needs=['2'], provides=['1'])
+    assert len(gossip.get_hook(timeline.hook_name)._registrations) == 2  # pylint: disable=protected-access
+
+    with pytest.raises(CannotResolveDependencies):
+        timeline.trigger()
+
+    timeline.get_hook().register_no_op(provides=['2'])
+    # Make sure "register_no_op" is not registred as "regular" callback
+    assert len(gossip.get_hook(timeline.hook_name)._registrations) == 2  # pylint: disable=protected-access
+    assert len(gossip.get_hook(timeline.hook_name)._empty_regisrations) == 1  # pylint: disable=protected-access
+
+    timeline.trigger()
+
+
+def test_unregister_no_op_remove_dependencies(timeline):
+    timeline.register(needs=['1'], provides=['0'])
+    timeline.register(needs=['2'], provides=['1'])
+
+    hook = timeline.get_hook()
+    registration = hook.register_no_op(provides=['2'])
+    timeline.trigger()
+
+    registration.unregister()
+    with pytest.raises(CannotResolveDependencies):
+        gossip.trigger(hook.full_name)
 
 
 @pytest.mark.parametrize('priority', [gossip.FIRST, gossip.DONT_CARE, gossip.LAST])
@@ -72,6 +135,7 @@ def test_policy_gets_reset():
 
 
 def test_unmet_dependencies(timeline):
+    # pylint: disable=unused-variable
     evt = timeline.register(needs=['a'])
     with pytest.raises(CannotResolveDependencies):
         timeline.trigger()